# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Hospital Management System API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /patients:
    get:
      operationId: listPatients
      tags: [patients]
      summary: List all patients
      parameters:
        - name: search
          in: query
          schema:
            type: string
        - name: status
          in: query
          schema:
            type: string
        - name: limit
          in: query
          schema:
            type: integer
        - name: offset
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Patients list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PatientList"
    post:
      operationId: createPatient
      tags: [patients]
      summary: Register a new patient
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PatientInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Patient"
  /patients/{id}:
    get:
      operationId: getPatient
      tags: [patients]
      summary: Get patient by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Patient
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Patient"
    put:
      operationId: updatePatient
      tags: [patients]
      summary: Update patient
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PatientInput"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Patient"
    delete:
      operationId: deletePatient
      tags: [patients]
      summary: Delete patient
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
  /patients/{id}/timeline:
    get:
      operationId: getPatientTimeline
      tags: [patients]
      summary: Get patient activity timeline
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Timeline
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Activity"

  # ─── Appointments ────────────────────────────────────────────
...
  /wards:
    get:
      operationId: listWards
      tags: [beds]
      summary: List all wards
      responses:
        "200":
          description: Wards
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Ward"
    post:
      operationId: createWard
      tags: [beds]
      summary: Create a ward
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/WardInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Ward"
  /beds:
    get:
      operationId: listBeds
      tags: [beds]
      summary: List all beds
      parameters:
        - name: wardId
          in: query
          schema:
            type: integer
        - name: status
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Beds
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Bed"
    post:
      operationId: createBed
      tags: [beds]
      summary: Create a bed
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BedInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Bed"
...
  /staff:
    get:
      operationId: listStaff
      tags: [staff]
      summary: List all staff members
      parameters:
        - name: role
          in: query
          schema:
            type: string
        - name: departmentId
          in: query
          schema:
            type: integer
        - name: search
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Staff list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/StaffMember"
    post:
      operationId: createStaff
      tags: [staff]
      summary: Add a staff member
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/StaffInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StaffMember"
  /staff/{id}:
    get:
      operationId: getStaffMember
      tags: [staff]
      summary: Get staff member by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Staff member
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StaffMember"
    put:
      operationId: updateStaff
      tags: [staff]
      summary: Update staff member
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/StaffInput"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StaffMember"
    delete:
      operationId: deleteStaff
      tags: [staff]
      summary: Delete staff member
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
components:
  schemas:
...
    SuccessResponse:
      type: object
      required: [success]
      properties:
        success:
          type: boolean
        message:
          type: string
...
    Activity:
      type: object
      required: [id, type, description, createdAt]
      properties:
        id:
          type: integer
        type:
          type: string
        description:
          type: string
        patientName:
          type: ["string", "null"]
        doctorName:
          type: ["string", "null"]
        createdAt:
          type: string

    # ─── Department ────────────────────────────
...
    Patient:
      type: object
      required: [id, patientId, name, email, gender, status, createdAt]
      properties:
        id:
          type: integer
        patientId:
          type: string
        name:
          type: string
        email:
          type: string
        phone:
          type: ["string", "null"]
        dateOfBirth:
          type: ["string", "null"]
        gender:
          type: string
        bloodGroup:
          type: ["string", "null"]
        address:
          type: ["string", "null"]
        allergies:
          type: ["string", "null"]
        emergencyContactName:
          type: ["string", "null"]
        emergencyContactPhone:
          type: ["string", "null"]
        insuranceProvider:
          type: ["string", "null"]
        insuranceNumber:
          type: ["string", "null"]
        status:
          type: string
        createdAt:
          type: string
    PatientInput:
      type: object
      required: [name, email, gender]
      properties:
        name:
          type: string
        email:
          type: string
        phone:
          type: string
        dateOfBirth:
          type: string
        gender:
          type: string
        bloodGroup:
          type: string
        address:
          type: string
        allergies:
          type: string
        emergencyContactName:
          type: string
        emergencyContactPhone:
          type: string
        insuranceProvider:
          type: string
        insuranceNumber:
          type: string
        status:
          type: string
    PatientList:
      type: object
      required: [patients, total]
      properties:
        patients:
          type: array
          items:
            $ref: "#/components/schemas/Patient"
        total:
          type: integer

    # ─── Appointment ───────────────────────────
...
    Ward:
      type: object
      required: [id, name, type, totalBeds, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        type:
          type: string
        totalBeds:
          type: integer
        availableBeds:
          type: integer
        createdAt:
          type: string
    WardInput:
      type: object
      required: [name, type, totalBeds]
      properties:
        name:
          type: string
        type:
          type: string
        totalBeds:
          type: integer
    Bed:
      type: object
      required: [id, wardId, bedNumber, status, createdAt]
      properties:
        id:
          type: integer
        wardId:
          type: integer
        wardName:
          type: ["string", "null"]
        bedNumber:
          type: string
        status:
          type: string
        patientId:
          type: ["integer", "null"]
        patientName:
          type: ["string", "null"]
        createdAt:
          type: string
    BedInput:
      type: object
      required: [wardId, bedNumber]
      properties:
        wardId:
          type: integer
        bedNumber:
          type: string
        status:
          type: string
...
    StaffMember:
      type: object
      required: [id, name, email, role, status, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
        phone:
          type: ["string", "null"]
        role:
          type: string
        departmentId:
          type: ["integer", "null"]
        departmentName:
          type: ["string", "null"]
        qualification:
          type: ["string", "null"]
        joinDate:
          type: ["string", "null"]
        salary:
          type: ["number", "null"]
        status:
          type: string
        createdAt:
          type: string
    StaffInput:
      type: object
      required: [name, email, role]
      properties:
        name:
          type: string
        email:
          type: string
        phone:
          type: string
        role:
          type: string
        departmentId:
          type: integer
        qualification:
          type: string
        joinDate:
          type: string
        salary:
          type: number
        status:
          type: string