# lib/api-zod/src/generated/types/staffInput.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Hospital Management System API
 * OpenAPI spec version: 0.1.0
 */

export interface StaffInput {
  name: string;
  email: string;
  phone?: string;
  role: string;
  departmentId?: number;
  qualification?: string;
  joinDate?: string;
  salary?: number;
  status?: string;
}