# lib/api-zod/src/generated/types/ward.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Hospital Management System API
 * OpenAPI spec version: 0.1.0
 */

export interface Ward {
  id: number;
  name: string;
  type: string;
  totalBeds: number;
  availableBeds?: number;
  createdAt: string;
}